#3268 Trust store of the socket factory can be based on the credentials keystore
