+++ opal-core/src/main/java/org/obiba/opal/core/service/security/OpalSocketFactoryProvider.java
@@ -1,6 +1,9 @@
 package org.obiba.opal.core.service.security;
 
 import org.obiba.magma.SocketFactoryProvider;
+import org.obiba.ssl.AnyTrustManager;
+import org.obiba.ssl.X509ExtendedKeyManagerImpl;
+import org.obiba.ssl.X509TrustManagerImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -9,7 +12,14 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import javax.net.SocketFactory;
-import javax.net.ssl.SSLSocketFactory;
+import javax.net.ssl.*;
+import java.security.KeyManagementException;
+import java.security.KeyStoreException;
+import java.security.NoSuchAlgorithmException;
+import java.security.UnrecoverableKeyException;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Component
 public class OpalSocketFactoryProvider implements SocketFactoryProvider {
@@ -23,18 +33,59 @@ public class OpalSocketFactoryProvider implements SocketFactoryProvider {
   @Qualifier("systemKeyStoreService")
   private SystemKeyStoreService systemKeyStoreService;
 
+  @Autowired
+  private CredentialsKeyStoreService credentialsKeyStoreService;
+
   private SocketFactory socketFactory;
 
   @Override
   public SocketFactory getSocketFactory() {
     if (socketFactory != null) return socketFactory;
     try {
-      socketFactory = systemKeyStoreService.getKeyStore().getSSLContext(allowInvalidCertificates).getSocketFactory();
+      socketFactory = getSSLContext(allowInvalidCertificates).getSocketFactory();
     } catch (Exception e) {
       log.error("Failed building a socket factory based on internal keystore", e);
       socketFactory = SSLSocketFactory.getDefault();
     }
     return socketFactory;
   }
+
+  private SSLContext getSSLContext(boolean allowInvalidCertificates) throws NoSuchAlgorithmException, KeyManagementException, KeyStoreException, UnrecoverableKeyException {
+    KeyManager[] keyManagers = getKeyManagers();
+    TrustManager[] trustManagers = getTrustManagers(allowInvalidCertificates);
+    SSLContext context = SSLContext.getInstance("SSL");
+    context.init(keyManagers, trustManagers, null);
+    return context;
+  }
+
+  /**
+   * Get a trust manager based on the credentials keystore (known certificates).
+   *
+   * @param allowInvalidCertificates
+   * @return
+   * @throws NoSuchAlgorithmException
+   * @throws KeyStoreException
+   */
+  private TrustManager[] getTrustManagers(boolean allowInvalidCertificates) throws NoSuchAlgorithmException, KeyStoreException {
+    if (allowInvalidCertificates) {
+      return new TrustManager[]{new AnyTrustManager()};
+    } else {
+      return new TrustManager[] {new X509TrustManagerImpl(credentialsKeyStoreService.getKeyStore(), false, true)};
+    }
+  }
+
+  /**
+   * Get a key manager based on the system keystore.
+   *
+   * @return
+   */
+  private KeyManager[] getKeyManagers() throws NoSuchAlgorithmException, UnrecoverableKeyException, KeyStoreException {
+    KeyManagerFactory kmFact = KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
+    kmFact.init(null, null);
+    List<KeyManager> kms = Arrays.stream(kmFact.getKeyManagers()).collect(Collectors.toList());
+    kms.add(new X509ExtendedKeyManagerImpl(systemKeyStoreService.getKeyStore()));
+    return kms.toArray(new KeyManager[kms.size()]);
+  }
+
 
 }
